Remove and/or relocate some FERC Form 1 data tests
- Removed some leftover infrastructure in the FERC 1 validation tests
  that was needed to check for unique record_id values in a number of
  columns -- these tests had already been migrated into dbt.
- Added some missing unique record_id checks in FERC 1 output tables.
- Moved the FERC 1 XBRL filter for freshest data test from validtion to
  integration tests... since it should work on both full and fast FERC
  XBRL derive SQLite DBs.

Closes #4283, #4282
