+++ dbt/models/ferc1/out_ferc1__yearly_all_plants/schema.yml
@@ -62,6 +62,8 @@ sources:
           - name: plant_hours_connected_while_generating
           - name: plant_type
           - name: record_id
+            data_tests:
+              - dbt_expectations.expect_column_values_to_be_unique
           - name: water_limited_capacity_mw
           - name: fuel_cost_per_mmbtu
           - name: fuel_type
