+++ dbt/models/ferc1/out_ferc1__yearly_small_plants_sched410/schema.yml
@@ -15,6 +15,8 @@ sources:
           - name: plant_id_pudl
           - name: plant_name_ferc1
           - name: record_id
+            data_tests:
+              - dbt_expectations.expect_column_values_to_be_unique
           - name: capacity_mw
           - name: capex_per_mw
           - name: capex_total
